Refactored layers and networks by splitting them into declaration in hpp and definitions in cpp.
