+++ AI/include/NeuralNetworks/Agents/A3C/base_a3c_agent.hpp
@@ -65,7 +65,6 @@ public:
 
   //////////////////////////////////////////////////////////////////////////////
   /// A2C/A3C standard methods
-
   /**
    *
    * @param actor_loss
